pyproject.toml
[project]
name = "elt-dvd-rental"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
